coauthor changes

+++ src/modules/article/article.controller.ts
@@ -22,6 +22,8 @@ import {
 } from '@nestjs/swagger';
 import { CreateArticleDto } from './dto/create-article.dto';
 import { JwtAuthGuard } from '../auth/jwt/jwt-auth.guard';
+import { UserRole } from '../user/enums/roles.enum';
+import { SendPaymentDto } from '../email/dto/send_payment.dto';
 
 @Controller('article')
 export class ArticleController {
@@ -31,7 +33,7 @@ export class ArticleController {
   @Get('allMy')
   @UseGuards(JwtAuthGuard)
   @ApiBearerAuth()
-  @ApiOperation({ summary: 'Get my articles' })
+  @ApiOperation({ summary: 'Get all my articles' })
   async getAllMyArticles(@Req() req: any) {
     return await this.articleService.getAllMy(req.user.id);
   }
@@ -43,6 +45,37 @@ export class ArticleController {
   @ApiOperation({ summary: 'Get my deleted articles' })
   async getAllMyDeletedArticles(@Req() req: any) {
     return await this.articleService.getAllMyDeleted(req.user.id);
+  }
+
+  @ApiTags('Articles for user')
+  @Post('sendArticlePayment')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  @ApiOperation({ summary: 'Send article payment file' })
+  @ApiConsumes('multipart/form-data')
+  @UseInterceptors(FileInterceptor('file'))
+  @ApiBody({
+    schema: {
+      type: 'object',
+      properties: {
+        file: {
+          type: 'string',
+          format: 'binary',
+        },
+        articleId: { type: 'number' },
+      },
+    },
+  })
+  async sendPayment(
+    @Req() req,
+    @Body() sendPaymentDto: SendPaymentDto,
+    @UploadedFile() file: Express.Multer.File,
+  ) {
+    if (!file) {
+      throw new BadRequestException('File not found');
+    }
+    sendPaymentDto.file = file;
+    return await this.articleService.sendPayment(sendPaymentDto, req.user.id);
   }
 
   @ApiTags('Articles for user')
@@ -64,6 +97,10 @@ export class ArticleController {
         text: { type: 'string' },
         category: { type: 'string' },
         coauthors: { type: 'string', example: 'Ryan Gosling, Tyler Durden' },
+        coauthorsEmails: {
+          type: 'string',
+          example: '@ryan.gosling@alatoo.edu.kg, @tylerdurden@email.com',
+        },
       },
       required: ['title', 'text'],
     },
@@ -81,7 +118,6 @@ export class ArticleController {
     article.file = file;
     Object.assign(article, createArticleDto);
     return await this.articleService.createArticle(req.user.id, article);
-    // console.log(file);
   }
 
   @ApiTags('Articles for admin')
@@ -114,15 +150,49 @@ export class ArticleController {
 
   @ApiTags('Articles for admin')
   @Patch('delete/:id')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   @ApiOperation({ summary: 'Delete article by id' })
-  async deleteArticle(@Param('id') id: number) {
+  async deleteArticle(@Param('id') id: number, @Req() req) {
+    if (req.user.role != UserRole.ADMIN) {
+      throw new BadRequestException('Only admin has permission to this action');
+    }
     return await this.articleService.deleteArticle(+id);
   }
 
   @ApiTags('Articles for admin')
   @Patch('restore/:id')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   @ApiOperation({ summary: 'Restore article by id' })
-  async restoreArticle(@Param('id') id: number) {
+  async restoreArticle(@Param('id') id: number, @Req() req) {
+    if (req.user.role != UserRole.ADMIN) {
+      throw new BadRequestException('Only admin has permission to this action');
+    }
     return await this.articleService.restoreArticle(+id);
+  }
+
+  @ApiTags('Articles for admin')
+  @Patch('approve/:id')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  @ApiOperation({ summary: 'Approve article by id' })
+  async approveArticle(@Param('id') id: number, @Req() req) {
+    if (req.user.role != UserRole.ADMIN) {
+      throw new BadRequestException('Only admin has permission to this action');
+    }
+    return await this.articleService.approveArticle(+id);
+  }
+
+  @ApiTags('Articles for admin')
+  @Patch('changeVisibility/:id')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
+  @ApiOperation({ summary: 'Make article public or private' })
+  async changeVisibility(@Param('id') id: number, @Req() req) {
+    if (req.user.role != UserRole.ADMIN) {
+      throw new BadRequestException('Only admin has permission to this action');
+    }
+    return await this.articleService.changeVisibility(+id);
   }
 }
